requirements.txt
functions-framework
flask
google-api-python-client
google-auth
google-cloud-secret-manager
openpyxl
pandas
pyarrow
python-dotenv
gspread
requests